PHPCS: WordAds widgets
