+++ modules/wordads/php/class-wordads-sidebar-widget.php
@@ -1,4 +1,9 @@
 <?php
+/**
+ * Widget for adding ads to a sidebar.
+ *
+ * @package Jetpack.
+ */
 
 /**
  * Widget for inserting an ad into your sidebar
@@ -7,10 +12,24 @@
  */
 class WordAds_Sidebar_Widget extends WP_Widget {
 
+	/**
+	 * Allowed tags.
+	 *
+	 * @var string[]
+	 */
 	private static $allowed_tags = array( 'mrec', 'wideskyscraper' );
-	private static $num_widgets  = 0;
 
-	function __construct() {
+	/**
+	 * Number of widgets.
+	 *
+	 * @var int
+	 */
+	private static $num_widgets = 0;
+
+	/**
+	 * WordAds_Sidebar_Widget constructor.
+	 */
+	public function __construct() {
 		parent::__construct(
 			'wordads_sidebar_widget',
 			/** This filter is documented in modules/widgets/facebook-likebox.php */
@@ -22,6 +41,14 @@ class WordAds_Sidebar_Widget extends WP_Widget {
 		);
 	}
 
+	/**
+	 * The Widget outputter.
+	 *
+	 * @param array $args Widget args.
+	 * @param array $instance The Widget instance.
+	 *
+	 * @return bool|void
+	 */
 	public function widget( $args, $instance ) {
 		global $wordads;
 		if ( $wordads->should_bail() ) {
@@ -36,7 +63,7 @@ class WordAds_Sidebar_Widget extends WP_Widget {
 		$about      = __( 'Advertisements', 'jetpack' );
 		$width      = WordAds::$ad_tag_ids[ $instance['unit'] ]['width'];
 		$height     = WordAds::$ad_tag_ids[ $instance['unit'] ]['height'];
-		$unit_id    = 1 == self::$num_widgets ? 3 : self::$num_widgets + 3; // 2nd belowpost is '4'
+		$unit_id    = 1 === self::$num_widgets ? 3 : self::$num_widgets + 3; // 2nd belowpost is '4'
 		$section_id = 0 === $wordads->params->blog_id ?
 			WORDADS_API_TEST_ID :
 			$wordads->params->blog_id . $unit_id;
@@ -44,9 +71,9 @@ class WordAds_Sidebar_Widget extends WP_Widget {
 		$snippet = '';
 		if ( $wordads->option( 'wordads_house', true ) ) {
 			$unit = 'mrec';
-			if ( 'leaderboard' == $instance['unit'] && ! $this->params->mobile_device ) {
+			if ( 'leaderboard' === $instance['unit'] && ! $this->params->mobile_device ) {
 				$unit = 'leaderboard';
-			} elseif ( 'wideskyscraper' == $instance['unit'] ) {
+			} elseif ( 'wideskyscraper' === $instance['unit'] ) {
 				$unit = 'widesky';
 			}
 
@@ -55,20 +82,27 @@ class WordAds_Sidebar_Widget extends WP_Widget {
 			$snippet = $wordads->get_ad_snippet( $section_id, $height, $width, 'widget' );
 		}
 
-		echo <<< HTML
+		?>
 		<div class="wpcnt">
 			<div class="wpa">
-				<span class="wpa-about">$about</span>
-				<div class="u {$instance['unit']}">
-					$snippet
+				<span class="wpa-about"><?php echo esc_html( $about ); ?></span>
+				<div class="u <?php echo esc_attr( $instance['unit'] ); ?>">
+					<?php echo $snippet; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped ?>
 				</div>
 			</div>
 		</div>
-HTML;
+		<?php
 	}
 
+	/**
+	 * The widget settings form.
+	 *
+	 * @param array $instance Widget instance.
+	 *
+	 * @return string|void
+	 */
 	public function form( $instance ) {
-		// ad unit type
+		// ad unit type.
 		if ( isset( $instance['unit'] ) ) {
 			$unit = $instance['unit'];
 		} else {
@@ -76,18 +110,18 @@ HTML;
 		}
 		?>
 		<p>
-			<label for="<?php echo esc_attr( $this->get_field_id( 'unit' ) ); ?>"><?php _e( 'Tag Dimensions:', 'jetpack' ); ?></label>
+			<label for="<?php echo esc_attr( $this->get_field_id( 'unit' ) ); ?>"><?php esc_html_e( 'Tag Dimensions:', 'jetpack' ); ?></label>
 			<select class="widefat" id="<?php echo esc_attr( $this->get_field_id( 'unit' ) ); ?>" name="<?php echo esc_attr( $this->get_field_name( 'unit' ) ); ?>">
 		<?php
 		foreach ( WordAds::$ad_tag_ids as $ad_unit => $properties ) {
-			if ( ! in_array( $ad_unit, self::$allowed_tags ) ) {
+			if ( ! in_array( $ad_unit, self::$allowed_tags, true ) ) {
 				continue;
 			}
 
 				$splits      = explode( '_', $properties['tag'] );
 				$unit_pretty = "{$splits[0]} {$splits[1]}";
 				$selected    = selected( $ad_unit, $unit, false );
-				echo "<option value='", esc_attr( $ad_unit ) ,"' ", $selected, '>', esc_html( $unit_pretty ) , '</option>';
+				echo "<option value='", esc_attr( $ad_unit ) ,"' ", esc_attr( $selected ), '>', esc_html( $unit_pretty ) , '</option>';
 		}
 		?>
 			</select>
@@ -95,10 +129,18 @@ HTML;
 		<?php
 	}
 
+	/**
+	 * The Widget updater.
+	 *
+	 * @param array $new_instance The revised instance.
+	 * @param array $old_instance Original instance.
+	 *
+	 * @return array
+	 */
 	public function update( $new_instance, $old_instance ) {
 		$instance = $old_instance;
 
-		if ( in_array( $new_instance['unit'], self::$allowed_tags ) ) {
+		if ( in_array( $new_instance['unit'], self::$allowed_tags, true ) ) {
 			$instance['unit'] = $new_instance['unit'];
 		} else {
 			$instance['unit'] = 'mrec';
@@ -107,35 +149,3 @@ HTML;
 		return $instance;
 	}
 }
-
-class WordAds_Ccpa_Do_Not_Sell_Link_Widget extends WP_Widget {
-	function __construct() {
-		parent::__construct(
-			'wordads_ccpa_do_not_sell_link_widget',
-			/** This filter is documented in modules/widgets/facebook-likebox.php */
-			apply_filters( 'jetpack_widget_name', __( 'Do Not Sell Link (CCPA)', 'jetpack' ) ),
-			array(
-				'description'                 => __( 'Inserts "Do Not Sell My Personal Information" link required by the California Consumer Privacy Act (CCPA)', 'jetpack' ),
-				'customize_selective_refresh' => true,
-			)
-		);
-	}
-
-	public function widget( $args, $instance ) {
-		echo $args['before_widget'];
-		echo do_shortcode( '[ccpa-do-not-sell-link]' );
-		echo $args['after_widget'];
-	}
-}
-
-function jetpack_wordads_widgets_init_callback() {
-	register_widget( 'WordAds_Sidebar_Widget' );
-
-	$ccpa_enabled = get_option( 'wordads_ccpa_enabled' );
-
-	if ( $ccpa_enabled ) {
-		register_widget( 'WordAds_Ccpa_Do_Not_Sell_Link_Widget' );
-	}
-}
-
-add_action( 'widgets_init', 'jetpack_wordads_widgets_init_callback' );
